[Breaking] Make methods in CompleteResult panic
Most of the time these methods will be unwrapped, as the
developer already knows how they set up the Recognizer and thus knows
the variant of the result *for sure*.

This change allows the library to give users detailed panic messages,
instead of the generic unwrap one.

Even when the variant is not known, needing an Option is extremely rare,
and if-let and match can be used in those cases.